==> SZZ/code_document/00_creating_new_directories.ipynb <==
import os

project = ""        # put the cloned repository name in the project, e.g. activemq, hbase, etc.
bugidentifier = ""  # put the bug identifier from the project, e.g. AMQ for activemq project, 
                    # HBASE for hbase project, etc.
userhome = os.path.expanduser('~')
analyze_dir = userhome + r"/DifferentDiffAlgorithms/SZZ/"

dirs = ['projects_analyses/',
        'projects_analyses/' + project + '/',
        'projects_analyses/' + project + '/01_bug_ids_extraction/',
        'projects_analyses/' + project + '/02_diff_extraction/',
        'projects_analyses/' + project + '/02_diff_extraction/01_modified_files/',
        'projects_analyses/' + project + '/02_diff_extraction/02_list_of_file_with_number_of_deletedlines/',
        'projects_analyses/' + project + '/02_diff_extraction/03_file_having_deletedlines/',
        'projects_analyses/' + project + '/02_diff_extraction/03_file_having_deletedlines/histogram/',
        'projects_analyses/' + project + '/02_diff_extraction/03_file_having_deletedlines/myers/',
        'projects_analyses/' + project + '/03_annotate/',
        'projects_analyses/' + project + '/03_annotate/01_annotated_files/',
        'projects_analyses/' + project + '/03_annotate/01_annotated_files/blame_histogram/',
        'projects_analyses/' + project + '/03_annotate/01_annotated_files/blame_histogram/csv/',
        'projects_analyses/' + project + '/03_annotate/01_annotated_files/blame_myers/',
        'projects_analyses/' + project + '/03_annotate/01_annotated_files/blame_myers/csv/',
        'projects_analyses/' + project + '/03_annotate/01_annotated_files/listof_diff_n_annotated_files/',
        'projects_analyses/' + project + '/03_annotate/02_diff-file_blame-file_comparison/',
        'projects_analyses/' + project + '/03_annotate/02_diff-file_blame-file_comparison/histogram_comparison/',
        'projects_analyses/' + project + '/03_annotate/02_diff-file_blame-file_comparison/myers_comparison/',
        'projects_analyses/' + project + '/04_affected_versions/',
        'projects_analyses/' + project + '/04_affected_versions/01_no_of_bugclass/',
        'projects_analyses/' + project + '/04_affected_versions/02_grouping_affected_non-affected_files/',
        'projects_analyses/' + project + '/05_validation/',
        'projects_analyses/' + project + '/05_validation/01_valid_buggyline/',
        'projects_analyses/' + project + '/05_validation/02_validfiles/',
        'projects_analyses/' + project + '/05_validation/03_validbugfixcommitid/',
        'projects_analyses/' + project + '/05_validation/04_validbugintrocommitid/',
        'projects_analyses/' + project + '/05_validation/05_stats/'
       ]

for z, newpath in enumerate(dirs):
    if not os.path.exists(analyze_dir + newpath):
        os.makedirs(analyze_dir + newpath)

with open (analyze_dir + "code_document/project_identity.txt", mode="wt", encoding="utf-8") as myfile:
    myfile.write('\n'.join([project, bugidentifier.upper()]))
print ("File project_identity.txt has been created")
