==> SZZ/code_document/06_comparing_buglineinoriginfile_vs_blamefiles.ipynb <==
# Importing libraries

import os, glob, csv, sys, re
from os import path
import pandas as pd

# Configure repository and directories

userhome = os.path.expanduser('~')
txt_file = open(userhome + r"/DifferentDiffAlgorithms/SZZ/code_document/project_identity.txt", "r")
pid = txt_file.read().split('\n')
project = pid[0]
bugidentifier = pid[1]
analyze_dir = userhome + r'/DifferentDiffAlgorithms/SZZ/projects_analyses/' + project + '/'

print ("Project name = %s" % project)
print ("Project key = %s" % bugidentifier)

# Loading the earliest affected versions dataset

affected_version = pd.read_csv(analyze_dir + '04_affected_versions/earliest_version.csv')
cols = ['bug_id','earliest_affected_version','date_release']
aff = affected_version[cols]

algorithms = ['histogram','myers']

colname = ['bug_id','bugfix_commitID','parent_id','filepath','diff_myers_file','diff_histogram_file','blame_myers_file','blame_histogram_file',
           '#deletions_myers','#deletions_histogram']
identified_files = pd.read_csv(analyze_dir + '04_affected_versions/affected_version_for_identified_files.csv')
identified_files

# Start comparing the bug-related lines with the line from the annotated files

# Histogram

blamehist_dirs = analyze_dir + '03_annotate/01_annotated_files/blame_histogram/csv/'
blamefile_hist = glob.glob(blamehist_dirs + '*')
print ("Found " + str(len(blamefile_hist)) + " files") 

diffhist_dirs = analyze_dir + '02_diff_extraction/03_file_having_deletedlines/histogram/'
difffile_hist = glob.glob(diffhist_dirs + '*')
print ("Found " + str(len(difffile_hist)) + " files") 

histdbfile = []
for num in range(0,len(identified_files)):
    sys.stdout.write('\rComparing bug-related lines and annotated files: %i' % (num+1) + ' out of %i' % (len(identified_files)))
    sys.stdout.flush()
    
    idfile = identified_files.iloc[num]
    blame = blamehist_dirs + idfile[7]
    
    if idfile[7] != "-":
    
        #Open csv file from result of git blame for <filename>.xxx
        fields = ['bugintro_commitid','line_number', 'code', 'date']
        blamefile = pd.read_csv(blame)
        blamefile = blamefile[fields]

        origin_code = []
        for n in range(0, len(blamefile)):
            temp = [blamefile.iloc[n][0], blamefile.iloc[n][1], str(blamefile.iloc[n][2]).replace('nan',''), blamefile.iloc[n][3]]
            origin_code.append(temp)

        #Open diff_file <filename>.xxx-xxxx_<diffalgorithm>buglines.diff vs blame_<filename>
        dfname = diffhist_dirs + idfile[5]    
        diff_file = open(dfname, encoding="utf8", errors='ignore')
        bugline = diff_file.read().split('\n')

        if bugline[len(bugline)-1] == '':
            for i in range(0,len(bugline)-1):
                bugline[i] = bugline[i][1:]
            del bugline[len(bugline)-1]
        else:
            for i in range(0,len(bugline)):
                bugline[i] = bugline[i][1:]

        bug_id = idfile[0]

        for vers in range(0,len(aff)):
            if bug_id == aff.iloc[vers][0]:
                version = aff.iloc[vers][1]
                daterelease = aff.iloc[vers][2]
                break;
            else:
                version = 'None'
                daterelease = 'None'

        num = 0
        blame_compare = []
        org_code = [el[2] for el in origin_code]
        for y, bugs in enumerate(bugline):
            if bugs in org_code[num:len(org_code)]:
                for x in range(num, len(origin_code)):                
                    codex = origin_code[x][2]
                    if bugs == codex:
                        if version != 'None':
                            # comparing the release date of affected version and the date the origin code was added
                            if origin_code[x][3] <= daterelease:
                                flag = 'bug-introducing change'
                            else:
                                flag = 'incorrect'
                        else:
                            flag = 'unknown'
                        temp = [bug_id, origin_code[x][0], origin_code[x][1], bugs, "True", origin_code[x][3], version, daterelease, flag]
                        blame_compare.append(temp)
                        num = origin_code[x][1]
                        break;                  
            else:
                temp = [bug_id, 'None', 'None', bugs, "False", 'None', version, daterelease, 'unknown']
                blame_compare.append(temp)
        
        fnames = ((idfile[3].split('/'))[-1:])[0]
        blamename = idfile[7]
        hist_name = 'diff_' + fnames + '_histogram-' + blamename
            
    else:
        blame_compare = []
        fnames = ((idfile[3].split('/'))[-1:])[0]
        blamename = idfile[7]
        hist_name = 'diff_' + fnames + '_histogram-' + blamename + '.csv'

    with open (analyze_dir + r'03_annotate/02_diff-file_blame-file_comparison/histogram_comparison/' + hist_name, 'w') as csvfile:
        writers = csv.writer(csvfile, delimiter = ",")
        header = ['bug_id', 'bugintro_commitid', 'line_number', 'buggy_code', 'does_bugline_from_originfile?', 'date_added_in_originfile', 'earliest_affected_version', 'version_release_date', 'bug_class']
        writers.writerow(header)
        for item in blame_compare:
            writers.writerow(item)
    
    histdbfile.append(hist_name)
xfile = pd.read_csv(analyze_dir + '04_affected_versions/affected_version_for_identified_files.csv')
xfile['histogram_diffnblame_file'] = histdbfile
print ('\nComparing bug-related lines are complete')

# Myers

blamemyers_dirs = analyze_dir + '03_annotate/01_annotated_files/blame_myers/csv/'
blamefile_myers = glob.glob(blamemyers_dirs + '*')
print ("Found " + str(len(blamefile_myers)) + " files") 

diffmyers_dirs = analyze_dir + '02_diff_extraction/03_file_having_deletedlines/myers/'
difffile_myers = glob.glob(diffmyers_dirs + '*')
print ("Found " + str(len(difffile_myers)) + " files") 

myersdbfile = []
for num in range(0,len(identified_files)):
    sys.stdout.write('\rComparing bug-related lines and annotated files: %i' % (num+1) + ' out of %i' % (len(identified_files)))
    sys.stdout.flush()
    
    idfile = identified_files.iloc[num]
    blame = blamemyers_dirs + idfile[6]
    
    if idfile[6] != "-":
    
        #Open csv file from result of git blame for <filename>.xxx
        fields = ['bugintro_commitid','line_number', 'code', 'date']
        blamefile = pd.read_csv(blame)
        blamefile = blamefile[fields]

        origin_code = []
        for n in range(0, len(blamefile)):
            temp = [blamefile.iloc[n][0], blamefile.iloc[n][1], str(blamefile.iloc[n][2]).replace('nan',''), blamefile.iloc[n][3]]
            origin_code.append(temp)

        #Open diff_file <filename>.xxx-xxxx_<diffalgorithm>buglines.diff vs blame_<filename>
        dfname = diffmyers_dirs + idfile[4]    
        diff_file = open(dfname, encoding="utf8", errors='ignore')
        bugline = diff_file.read().split('\n')

        if bugline[len(bugline)-1] == '':
            for i in range(0,len(bugline)-1):
                bugline[i] = bugline[i][1:]
            del bugline[len(bugline)-1]
        else:
            for i in range(0,len(bugline)):
                bugline[i] = bugline[i][1:]

        bug_id = idfile[0]

        for vers in range(0,len(aff)):
            if bug_id == aff.iloc[vers][0]:
                version = aff.iloc[vers][1]
                daterelease = aff.iloc[vers][2]
                break;
            else:
                version = 'None'
                daterelease = 'None'

        num = 0
        blame_compare = []
        org_code = [el[2] for el in origin_code]
        for y, bugs in enumerate(bugline):
            if bugs in org_code[num:len(org_code)]:
                for x in range(num, len(origin_code)):                
                    codex = origin_code[x][2]
                    if bugs == codex:
                        if version != 'None':
                            # comparing the release date of affected version and the date the origin code was added
                            if origin_code[x][3] <= daterelease:
                                flag = 'bug-introducing change'
                            else:
                                flag = 'incorrect'
                        else:
                            flag = 'unknown'
                        temp = [bug_id, origin_code[x][0], origin_code[x][1], bugs, "True", origin_code[x][3], version, daterelease, flag]
                        blame_compare.append(temp)
                        num = origin_code[x][1]
                        break;                  
            else:
                temp = [bug_id, 'None', 'None', bugs, "False", 'None', version, daterelease, 'unknown']
                blame_compare.append(temp)
                
        fnames = ((idfile[3].split('/'))[-1:])[0] 
        blamename = idfile[6]
        myers_name = 'diff_' + fnames + '_myers-' + blamename
    
    else:
        blame_compare = []
        fnames = ((idfile[3].split('/'))[-1:])[0] 
        blamename = idfile[6]
        myers_name = 'diff_' + fnames + '_myers-' + blamename + '.csv'

    with open (analyze_dir + r'03_annotate/02_diff-file_blame-file_comparison/myers_comparison/' + myers_name, 'w') as csvfile:
        writers = csv.writer(csvfile, delimiter = ",")
        header = ['bug_id', 'bugintro_commitid', 'line_number', 'buggy_code', 'does_bugline_from_originfile?', 'date_added_in_originfile', 'earliest_affected_version', 'version_release_date', 'bug_class']
        writers.writerow(header)
        for item in blame_compare:
            writers.writerow(item)

    myersdbfile.append(myers_name)
        
xfile['myers_diffnblame_file'] = myersdbfile   
print ('\nComparing buglines are complete')

colname = ['bug_id','bugfix_commitID','parent_id','filepath','diff_myers_file','diff_histogram_file','blame_myers_file',
           'blame_histogram_file','myers_diffnblame_file','histogram_diffnblame_file','#deletions_myers',
           '#deletions_histogram','earliest_affected_version','date_release']
xfile = xfile[colname]
xfile.to_csv(analyze_dir + '03_annotate/02_diff-file_blame-file_comparison/diffnblame_comparison_files.csv')
xfile

# Comparing the number of bug class of each algorithm

algorithm = ['histogram','myers']

#set the filepath of the filenames
filepath = analyze_dir + '03_annotate/02_diff-file_blame-file_comparison/'

result = []

for ii, alg in enumerate(algorithm):
    bug_intro_qty = incorrect_qty = unknown_qty = sumln = 0
    i = 1
    for fname in glob.iglob(filepath + alg + '_comparison/*', recursive=True):
        sys.stdout.write('\r%i ' %(ii+1) + 'Counting the number of bug-class in file %i' % i)
        sys.stdout.flush()
        i += 1

        dframe = pd.read_csv(fname, low_memory=False)
        op = open(fname, encoding="utf8", errors='ignore')
        lines = op.read().split('\n')
        if "-- /dev" not in lines[1]:
            sumln = (len(lines)-2) + sumln
            bug_intro_qty = dframe['bug_class'][dframe['bug_class'] == 'bug-introducing change'].count() + bug_intro_qty
            incorrect_qty = dframe['bug_class'][dframe['bug_class'] == 'incorrect'].count() + incorrect_qty
            unknown_qty = dframe['bug_class'][dframe['bug_class'] == 'unknown'].count() + unknown_qty
        else:
            sumln = (len(lines)-3) + sumln
            pass
    
    result.append([bug_intro_qty, incorrect_qty, unknown_qty, sumln])
print ('\nCounting the number of bug-introducing change is complete')

result_matrix = [list(i) for i in zip(*result)]
label = ['#bug-introducing_change', '#incorrect', '#unknown', 'total']
for s, lbl in enumerate(label):
    result_matrix[s].insert(0, lbl)
result_matrix

with open(analyze_dir + '04_affected_versions/01_no_of_bugclass/quantity_of_bug_class.csv','w') as myfile:
    header = ['class', 'histogram', 'myers']
    writers = csv.writer(myfile)
    writers.writerow(header)
    for item in result_matrix:
        writers.writerow(item)

df_bugclass = pd.read_csv(analyze_dir + '04_affected_versions/01_no_of_bugclass/quantity_of_bug_class.csv')
df_bugclass = df_bugclass[header]
df_bugclass
