==> SZZ/code_document/07_grouping_the_affected_files.ipynb <==
# Importing libraries

import os, glob, csv, sys, re, operator
from os import path
import pandas as pd

# Configure repository and directories

userhome = os.path.expanduser('~')
txt_file = open(userhome + r"/DifferentDiffAlgorithms/SZZ/code_document/project_identity.txt", "r")
pid = txt_file.read().split('\n')
project = pid[0]
bugidentifier = pid[1]
analyze_dir = userhome + r'/DifferentDiffAlgorithms/SZZ/projects_analyses/' + project + '/'

print ("Project name = %s" % project)
print ("Project key = %s" % bugidentifier)

# Loading comparison data between bugline and blame

compare_dirs = analyze_dir + '03_annotate/02_diff-file_blame-file_comparison/'

dirs = []
for root, folder, fileList in os.walk(compare_dirs):
    if folder != []:
        dirs.append(folder)
dirs.sort()
dirs = dirs[0]
dirs

# Load dataset of diff and blame file

colname = ['bug_id','bugfix_commitID','parent_id','filepath','diff_myers_file','diff_histogram_file','blame_myers_file',
           'blame_histogram_file','myers_diffnblame_file','histogram_diffnblame_file','#deletions_myers',
           '#deletions_histogram','earliest_affected_version','date_release']
df_diffblame = pd.read_csv(compare_dirs + 'diffnblame_comparison_files.csv')
df_diffblame = df_diffblame[colname]
df_diffblame

# Reading files from comparison dataset

#Reading files from comparison
for n in range(0,len(dirs)):
    result = []
    for diff in range(0,len(df_diffblame)):        
        sys.stdout.write("\r{} Reading file no: {}".format((n+1),(diff+1)))
        sys.stdout.flush()
        dt = df_diffblame.iloc[diff]
        file = compare_dirs + dirs[n] + "/" + dt[8+n]

        res = []
        with open(file,'r') as myfile:
            csv_reader = csv.reader(myfile, delimiter=',')
            for row in csv_reader:
                res.append(row[8])
        del res[0]
        
        filename = ((dt[3].split('/'))[-1:])[0] 

        num_validbug = 0
        if ('bug-introducing change') not in res:
            tmp = [dt[0],dt[1],dt[2],dt[3],dt[4],dt[5],dt[6],dt[7],dt[8],dt[9],dt[10],dt[11],dt[12],dt[13],
                   num_validbug,"Does not affect the version"]
        else:
            df = pd.read_csv(file)
            num_validbug = df['bug_class'][df['bug_class'] == 'bug-introducing change'].count()
            tmp = [dt[0],dt[1],dt[2],dt[3],dt[4],dt[5],dt[6],dt[7],dt[8],dt[9],dt[10],dt[11],dt[12],dt[13],
                   num_validbug,"Affect the version"]
        result.append(tmp)
    
    algorithm = re.search("(\w+)_comparison", dirs[n])
    algorithm = algorithm.group(1)

    with open(analyze_dir + "04_affected_versions/02_grouping_affected_non-affected_files/" + algorithm + "_affnonaff_files.csv",'w') as csvfile:
        fields = colname + [algorithm + '_#validbugline',algorithm + '_affect_version?']
        writers = csv.writer(csvfile, delimiter=",")
        writers.writerow(fields)
        for item in result:
            writers.writerow(item)
            
print ("\nThe process is complete")

cname = ['bug_id','bugfix_commitID','parent_id','filepath','diff_histogram_file','blame_histogram_file','histogram_diffnblame_file',
         '#deletions_histogram','earliest_affected_version','date_release','histogram_#validbugline',
         'histogram_affect_version?']
df_affhist = pd.read_csv(analyze_dir + "04_affected_versions/02_grouping_affected_non-affected_files/histogram_affnonaff_files.csv")
df_affhist = df_affhist[cname]
histafffile = df_affhist[cname][df_affhist[cname]['histogram_affect_version?'] == 'Affect the version']
histafffile.to_csv(analyze_dir + "04_affected_versions/02_grouping_affected_non-affected_files/histogram_affected_files.csv",index=False)
histafffile

cname = ['bug_id','bugfix_commitID','parent_id','filepath','diff_myers_file','blame_myers_file','myers_diffnblame_file',
         '#deletions_myers','earliest_affected_version','date_release','myers_#validbugline','myers_affect_version?']
df_affmyers = pd.read_csv(analyze_dir + "04_affected_versions/02_grouping_affected_non-affected_files/myers_affnonaff_files.csv")
df_affmyers = df_affmyers[cname]
myersafffile = df_affmyers[cname][df_affmyers[cname]['myers_affect_version?'] == 'Affect the version']
myersafffile.to_csv(analyze_dir + "04_affected_versions/02_grouping_affected_non-affected_files/myers_affected_files.csv",index=False)
myersafffile

# Counting affected and non-affected files

df = [df_affhist,df_affmyers]
algorithms = ['histogram','myers']

no_of_files = []
for z, dframe in enumerate(df):
    affnum = dframe[algorithms[z] + '_affect_version?'][dframe[algorithms[z] + '_affect_version?'] == 'Does not affect the version'].count()
    nonaffnum = dframe[algorithms[z] + '_affect_version?'][dframe[algorithms[z] + '_affect_version?'] == 'Affect the version'].count()
    summ = [algorithms[z],affnum,nonaffnum]
    no_of_files.append(summ)

with open(analyze_dir + "04_affected_versions/02_grouping_affected_non-affected_files/totalnumber_of_affandnonaff_files.csv",'w') as file:
    header = ['algorithm','#non-affected_files','#affected_files']
    writers = csv.writer(file,delimiter=",")
    writers.writerow(header)
    for ii in no_of_files:
        writers.writerow(ii)
print ("File 'totalnumber_of_affandnonaff_files.csv' has been created")

aff_file = pd.read_csv(analyze_dir + "04_affected_versions/02_grouping_affected_non-affected_files/totalnumber_of_affandnonaff_files.csv")
aff_file
