==> SZZ/code_document/10_counting_number_of_validfiles_and_bugfixcommitid_having_different_locationofdiff.ipynb <==
# Import libraries

import os, glob, csv, sys, re, math
import pandas as pd
from os import path

# Configure directories

userhome = os.path.expanduser('~')
txt_file = open(userhome + r"/DifferentDiffAlgorithms/SZZ/code_document/project_identity.txt", "r")
pid = txt_file.read().split('\n')
project = pid[0]
bugidentifier = pid[1]
proj = project.upper()
analyze_dir = userhome + r'/DifferentDiffAlgorithms/SZZ/projects_analyses/' + project + '/'
compare_dir = analyze_dir + "02_diff_extraction/03_file_having_deletedlines/"

print ("Project name = %s" % project)
print ("Project key = %s" % bugidentifier)

# Load the dataset

#Load comparison dataset
diffblame_myers = pd.read_csv(analyze_dir + "05_validation/02_validfiles/myers_valid_files.csv")
diffblame_hist = pd.read_csv(analyze_dir + "05_validation/02_validfiles/histogram_valid_files.csv")
diffblame_myers

Merge 2 dataset

diffblame = diffblame_myers.merge(diffblame_hist, on=['bug_id','bugfix_commitID','parent_id','filepath'], how='outer')
diffblame.fillna(0, inplace=True)
diffblame = diffblame.reset_index(drop=True)
diffblame = diffblame[['bug_id','bugfix_commitID','parent_id','filepath','diff_myers_file','diff_histogram_file',
                       'blame_myers_file','blame_histogram_file','myers_diffnblame_file','histogram_diffnblame_file',
                       'myers_#validbugline','histogram_#validbugline','myers_affect_version?','histogram_affect_version?']]
diffblame.to_csv(analyze_dir + "05_validation/02_validfiles/validfiles_combination.csv", index=False)
diffblame

data_equals = []
data_unequals = []
for dd in range (0,len(diffblame)):
    dbl = diffblame.iloc[dd]
    if (dbl[4] != 0) and (dbl[5] != 0):
        data_equals.append(dd)
    else:
        if dbl[4] == 0:
            data_unequals.append([dd,"histogram/",dbl[5]])
        else:
            data_unequals.append([dd,"myers/",dbl[4]])

# Start Comparing Files

comparison_files = []
for m, number in enumerate(data_equals):
    sys.stdout.write("\rComparing file no.: {} / {}".format((m+1),len(data_equals)))
    sys.stdout.flush()
    
    ff = diffblame.iloc[number]
    mfiles = compare_dir + "myers/" + ff[4]
    hfiles = compare_dir + "histogram/" + ff[5]
    
    hist_file = open(hfiles, encoding="utf8", errors='ignore')
    bugline_hist = hist_file.read().split('\n')

    if bugline_hist[len(bugline_hist)-1] == '':
        for i in range(0,len(bugline_hist)-1):
            bugline_hist[i] = bugline_hist[i][1:]
        del bugline_hist[len(bugline_hist)-1]
    else:
        for i in range(0,len(bugline_hist)):
            bugline_hist[i] = bugline_hist[i][1:]
    
    myers_file = open(mfiles, encoding="utf8", errors='ignore')
    bugline_myers = myers_file.read().split('\n')

    if bugline_myers[len(bugline_myers)-1] == '':
        for i in range(0,len(bugline_myers)-1):
            bugline_myers[i] = bugline_myers[i][1:]
        del bugline_myers[len(bugline_myers)-1]
    else:
        for i in range(0,len(bugline_myers)):
            bugline_myers[i] = bugline_myers[i][1:]
            
    if len(bugline_hist) > len(bugline_myers):
        x = len(bugline_myers)
    else:
        x = len(bugline_hist)
    
    sum_line = 0
    for b in range(0,x):        
        if bugline_hist[b] != bugline_myers[b]:
            sum_line = sum_line + 1
        else:
            sum_line = sum_line
    sum_line = sum_line + (abs(len(bugline_hist)-len(bugline_myers)))
    fnames = ((ff[3].split('/'))[-1:])[0] 
    temp = [number, ff[0], ff[1], ff[2], ff[3], fnames, sum_line]
    
    comparison_files.append(temp)
    
print ("\nThe process is complete")

for un, num in enumerate(data_unequals):
    sys.stdout.write("\rComparing file no.: {} / {}".format((un+1),len(data_unequals)))
    sys.stdout.flush()
    
    nn = num[0]
    file = compare_dir + num[1] + num[2]
    un_file = open(file, encoding="utf8", errors='ignore')        
    bugline_unfile = un_file.read().split('\n')

    if bugline_unfile[len(bugline_unfile)-1] == '':
        for i in range(0,len(bugline_unfile)-1):
            bugline_unfile[i] = bugline_unfile[i][1:]
        del bugline_unfile[len(bugline_unfile)-1]
    else:
        for i in range(0,len(bugline_unfile)):
            bugline_unfile[i] = bugline_unfile[i][1:]
            
    sum_line = len(bugline_unfile)
    zz = diffblame.iloc[nn]
    fnames = ((zz[3].split('/'))[-1:])[0] 
    temp = [number, zz[0], zz[1], zz[2], zz[3], fnames, sum_line]
    
    comparison_files.append(temp)
    
print ("\nThe process is complete")

with open(analyze_dir + "05_validation/01_valid_buggyline/number_of_difflocation_of_valid_deletedlines.csv",'w') as csvfile:
    header = ['data_no','bug_id','bugfix_commitid','parent_id','filepath','filename','#different_location_valid_deletedline']
    writers = csv.writer(csvfile, delimiter=",")
    writers.writerow(header)
    for item in comparison_files:
        writers.writerow(item)
print("The file has been created")

diffloc = pd.read_csv(analyze_dir + "05_validation/01_valid_buggyline/number_of_difflocation_of_valid_deletedlines.csv")
diffloc

file_withdiffloc = diffloc[header][diffloc[header]['#different_location_valid_deletedline'] != 0]
file_withdiffloc.to_csv(analyze_dir + '05_validation/02_validfiles/validfiles_with_different_locofbugline.csv')
file_withdiffloc

#grouping the bug-fix commit id
bugfixcid = diffloc.groupby('bugfix_commitid', as_index=False).agg({"#different_location_valid_deletedline":"sum"})
cols = ['bugfix_commitid','#different_location_valid_deletedline']
diffloc_bugfixcid = bugfixcid[cols][bugfixcid[cols]['#different_location_valid_deletedline'] != 0]
diffloc_bugfixcid.to_csv(analyze_dir + '05_validation/03_validbugfixcommitid/validbugfixcid_with_different_locofbugline.csv')
diffloc_bugfixcid

# Summary of the different location in file-level

difffile = len(file_withdiffloc)
samefile = len(diffloc)-len(file_withdiffloc)
totalfile = len(diffloc)
percentfile = (difffile/totalfile)*100

print ("Summary of file based on location")
print ("==================================")
print ("Project name : " + proj)
print ("Number of files that have different bugline : %i"%(difffile))
print ("Number of files that have same bugline : %i" %(samefile))
print ("Total number of files : %i" %(totalfile))
print ("Percentage of different : {0:.2f}%".format(percentfile))

# Summary of the different location in commit-level

diffcommit = len(diffloc_bugfixcid)
samecommit = len(bugfixcid)-len(diffloc_bugfixcid)
totalcommit = len(bugfixcid)
percentcommit = (diffcommit/totalcommit)*100

print ("Summary of commit based on location")
print ("==================================")
print ("Project name : " + proj)
print ("Number of commit id that have different bugline : %i"%(diffcommit))
print ("Number of commit id that have same bugline : %i" %(samecommit))
print ("Total number of commit id : %i" %(totalcommit))
print ("Percentage of different : {0:.2f}%".format(percentcommit))

#save in csv file
sum_data = [[proj,'file-level',samefile,difffile,percentfile,totalfile],
            [proj,'commit-level',samecommit,diffcommit,percentcommit,totalcommit]
           ]
with open (analyze_dir + "05_validation/05_stats/summary_of_difflocation_in_filencommit.csv",'w') as mycsv:
    head = ['project_name','level','#same','#different','percentage','totalnumber']
    writers = csv.writer(mycsv, delimiter=',')
    writers.writerow(head)
    for item in sum_data:
        writers.writerow(item)

csvfile = pd.read_csv(analyze_dir + "05_validation/05_stats/summary_of_difflocation_in_filencommit.csv")
csvfile
